+++ examples/data-sources/iosxe_crypto_pki/data-source.tf
@@ -1,0 +1,2 @@
+data "iosxe_crypto_pki" "example" {
+}