+++ examples/resources/iosxe_crypto_pki/import.sh
@@ -1,0 +1,1 @@
+terraform import iosxe_crypto_pki.example "Cisco-IOS-XE-native:native/crypto/Cisco-IOS-XE-crypto:pki"