+++ examples/resources/iosxe_crypto_pki/resource.tf
@@ -1,0 +1,9 @@
+resource "iosxe_crypto_pki" "example" {
+  trustpoints = [
+    {
+      id                = "trustpoint1"
+      enrollment_pkcs12 = true
+      revocation_check  = ["none"]
+    }
+  ]
+}